Move POSItem models to their own files
